Added check to see if ofac_sdns exists before dropping the table

+++ db/migrate/20140715173240_drop_ofac_sdns.rb
@@ -1,5 +1,5 @@
 class DropOfacSdns < ActiveRecord::Migration
   def change
-    drop_table :ofac_sdns
+    drop_table :ofac_sdns if ActiveRecord::Base.connection.table_exists? 'ofac_sdns'
   end
 end
